Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@5498 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ apple2/src/Disk.java
@@ -173,7 +173,7 @@ public class Disk
                         isValidFileType(sector[ce+2]) &&
                         isValidFileName(sector,ce+3))
                     {
-                        if (sector[ce+31] == 0xA0)
+                        if (sector[ce+31] == (byte)0xA0)
                         {
                             ++penultimateSpace;
                         }
